requirements.txt
opencv-python
numpy
mediapipe
tensorflow
Pillow